requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.3.0
charset-normalizer==3.0.1
cycler==0.11.0
Cython==0.29.33
easydict==1.10
flatbuffers==1.12
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.4
importlib-metadata==6.0.0
Keras==2.4.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.1.2.30
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.4.1
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.3
tensorflow-estimator==2.5.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.12.1
yolo34py==0.2
zipp==3.12.0
